changes in App, interface and Model

+++ src/main/java/uvg/edu/IBlender.java
@@ -15,7 +15,7 @@ public interface IBlender {
      **/
     public void fillBlender(String prod, float ml);
 
-    public int actualCapacity();
+    public float actualCapacity();
 
     public void fillBlender(String prod);
